Make database name valid in identifier context without quotes
If someone is programatically interacting with the database (e.g., via
ALTER DATABASE or similar), the dashes in the current identifier make
it so that the identifier must be quoted (i.e., ALTER DATABASE "$name"
rather than just allowing ALTER DATABASE $name). This is valid, but
the identifier is mostly to make it clear that it is a buildpack db,
so a fancy name is not necessary. Simplify this to allow using the
name without quotes.