Specify UTF-8 as default encoding for initdb
Currently, the locale is unset as part of this buildpack, so the
default encoding is `SQL_ASCII`, which is different to the add-on
default encoding of `UTF8`.

This commit specifies `UTF8` as the encoding for the template database
by passing it to `initdb`'s `-E` switch.

Fixes https://github.com/heroku/heroku-buildpack-ci-postgresql/issues/2